migrated config validator tests to logging harness
