+++ test/with/logging.go
@@ -1,4 +1,4 @@
-package test
+package with
 
 import (
 	"github.com/orbs-network/scribe/log"
@@ -15,7 +15,7 @@ func (h *LoggingHarness) AllowErrorsMatching(pattern string) {
 	h.testOutput.AllowErrorsMatching(pattern)
 }
 
-func WithLogger(tb testing.TB, f func(harness *LoggingHarness)) {
+func Logging(tb testing.TB, f func(harness *LoggingHarness)) {
 	testOutput := log.NewTestOutput(tb, log.NewHumanReadableFormatter())
 	h := &LoggingHarness{
 		Logger:     log.GetLogger().WithOutput(testOutput),
@@ -26,3 +26,18 @@ func WithLogger(tb testing.TB, f func(harness *LoggingHarness)) {
 	f(h)
 	RequireNoUnexpectedErrors(tb, testOutput)
 }
+
+type Fataler interface {
+	Fatal(args ...interface{})
+}
+
+type ErrorTracker interface {
+	HasErrors() bool
+}
+
+func RequireNoUnexpectedErrors(f Fataler, errorTracker ErrorTracker) {
+	if errorTracker.HasErrors() {
+		f.Fatal("Test failed; encountered unexpected errors")
+	}
+}
+
